bot/requirements.txt
beautifulsoup4==4.12.2
feedparser==6.0.10
requests==2.31.0
urllib3==2.0.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
tweepy==4.14.0



